Move towards using [unchanged_at_extended]. See msg.
The issue with the previous approach is because we need to be
conservative. This in particular means that if we have both
[unchanged_at locs s1 s2] and [unchanged_except locs s1 s2], we **DO
NOT** have [unchanged_except Nil s1 s2], even thought it might seem
like we should.

An example that demonstrates this is if we assume that XMM1 is not
disjoint from RAX. This is a conservative assumption and is correct
and sound. However, at this point, we face our conundrum. Assume the
[locs] above is the singleton set [RAX]. Since [unchanged_at]
looks **only** at the locations within the [locs] and
[unchanged_except] looks at locations which are _strictly_ disjoint
from the [locs], the [XMM1] is not talked about whatsoever. This means
that with this definition, we have a situation where the [XMM1]
register _could_ have different values, and thus this is a
counter-example.

We fix this by extending [unchanged_at], to get
[unchanged_at_extended]. This new extended definition takes into
account everything that is not provably disjoint. This means that it
is more difficult to prove, but it provides much more if it were to
exist in the precondition. Additionally, it properly encodes the
notion of conservatively equal. Thus, with this definition, we can
now (trivially) show that [unchanged_at_extended locs s1 s2] and
[unchanged_except locs s1 s2] implies [unchanged_except Nil s1 s2].

It does throw some extra stuff into the picture; in particular, we can
no longer as easily show the behavior upon two operations that are
commuted. We need to introduce a new notion of
[unchanged_upon_both_non_disjoint] which allows one to talk about two
different sets of access locations; and is true only if their
intersection is maintained across both states.

It remains to be seen how much extra effort this new approach will
take up for proving things from [untainted_eval_ins] though.