Some progress towards [unchanged_upon_both_non_disjoint]. See msg.
In order to prove [unchanged_upon_both_non_disjoint], we need to show
that our notion of [disjoint_access_locations] is _actually_
conservative. On the way to showing this, we need to additionally
introduce the notion of an [access_location] that is [always_constant]
since otherwise, it messes with the way the lists work and how
disjoint-ness works.