Handle memory locations more precisely. See msg.
For example, the following instruction actually _reads_ from RAX and
writes to some location in memory. Previously this would've been
considered to not really be a read from RAX but we were relying upon
the conservative behavior of the [disjoint] operation to be correct;
now it is baked into the definition of the read-write sets.

```
     mov [rax + 5], 10
```