Start work on non-eqtype permutation finding. See msg.
We are not going to make the [ins] into an [eqtype], since it
internally contains functions. Instead, a function equivalent to the
[eq_code] in the new code will be exposed by Chris, which will allow
one to know if the evaluation of two [code]s is the same. Since we do
_not_ care about the actual instructions but we care only about their
behavior, this should be sufficient. A small parts of the final proof
breaks due to this change, but this should be fixed up in the upcoming
commits.