Allow more things to move around! See msg.
Previously, we only allowed things to exchange amongst each other if
their rw-sets satisfied the following conditions: (1) their reads
don't coincide with the other's writes, and (2) their writes don't
coincide.

This is quite restrictive, since if two instructions were to perform a
write of a constant into a location, then it is actually safe to do
exchange them. With this commit, we can now check for this sort of
behavior and allow this sort of exchange.

In particular, we now check if two things have the exact same value
written. If they do so, then those values can be safely ignored when
comparing write-write set disjointness.

The concrete effect of this change is that now we are able to exchange
instructions that havoc the flags, as long as both instructions havoc
the flags.

---

This upgrade is a significant step forwards towards being able to
support the full level of instr-reordering which allows one to
exchange even entire blocks of code, even if individual instructions
themselves cannot be exchanged!

---

One MASSIVE caveat however is that this breaks our sanity-check for
instruction exchanges. In particular, if an instructions write values
are disjoint, it will disallow it due to non-same constant-write
values. To bring back the more forgiving behavior, it will require
examining and expanding the definition of [rw_exchange_allowed] to
allow more cases.