1st pass impl of new (more precise) exchange check. See msg.
A few commits ago, we had an issue with instructions that had
differing constant-writes. This fixes the issue. However, this breaks
some of the proofs that had been written them. This commit fixes some
of these proofs, but also introduces a few [admit]s which will be
fixed in future commits.