Remove dependence on any properties of equality on [ins]. See msg.
This change brings about two things:
+ [eq_code] -> [eq_ins]
+ Removal of dependence on the properties of the [eq_*]

The first is fairly obvious. We just reduce the scope of things that
Vale needs to provide. The second is a little more interesting.

We previously used to require knowing that [eq_*] meant
   (1) evaluations are the same, and
   (2) they should still remain exchange-able

We no longer have these requirements. In particular, we have _no_
requirements from [eq_ins] other than it being a boolean function that
gives _some_ notion of equivalence between instructions. We work with
this notion, without needing any further information from it.

The way this transformer works now is that given two [codes] objects,
say [c1] and [c2], it generates a new [codes] object [c_gen] that
consists _solely_ of a reordering of [c1], such that we will always
have [eq_codes c_gen c2]. Additionally, we obtain the property that
evaluating over [c1] leads to the same behavior as evaluating over
[c_gen] (up to [equiv_states]).

Next steps are to make the [c_gen] generation actually perform a
_deep-dive_ into [c1] rather than performing only the outermost layer.