Allow more swaps than were allowed before :) See msg.
Some lemmas broke due to this change and will be updated in upcoming
commits.

We now allow entire blocks of code to move past each other, as long as
they satisfy the r/w requirements. We explicitly DO NOT perform
block/unblock, and leave that to the user of this transformation to
do. However, if it is possible to move two [code]s objects past each
other, this will allow it.

At the moment, this is fairly limited, since we don't have very
precise definition for taking multiple [rw_set]s in series, but once
we have them, this should allow us to move larger chunks of code
across.