Support "diving into" blocks to find code. See msg.
Previously, we required both the original code and the hint to have an
equivalent set of blocks. Now instead, we allow the original code to
be deeper. The hint needs to be shallower and is used to decide how to
"dive into" the original code to find the relevant code.

The reason it is ok to not "break apart" the hint when diving in is
because the expected use case will only be a logical reordering of the
code, and not an arbitrary reordering. In this case, either we will
have the necessary blocks already, or we want to be at the level of
instructions, but in that case, we already will have those at the
hint.