New style for finding transformations. See msg.
Previously the part of the code that found the transformation was part
of the code that needed to be verified, with proofs written directly
upon it. Now instead, we allow for an unverified function to find a
transformation for us, and then we simply prove the part that performs
a suggested transformation. In particular, this code that finds a
transformation for us is now [irreducible] and nothing is known about
it, other than the fact that it returns an object that contains
[possibly transformation_hints]. This means that no proof relies upon
this function, but instead, we have the main proof rely upon
[perform_transformation_with_hint] and
[perform_transformations_with_hint].

There are numerous advantages to this approach:
(1) Proofs no longer need to be updated as we try different heuristics
    for performing the unblocking.
(2) Once we have a "default" [find_transformation_hints], we are still
    not limited by it, since someone could easily come along and write
    their own [find_transformation_hints] (or even produce their own
    [transformation_hints] object manually) and pass it into the
    framework to perform the transformation.

The disadvantages to this approach are fairly minimal. In particular,
we can no longer find and perform the transformation at the same time,
but instead must do it separately. This should not be an issue, since
earlier too, we were (effectively) doing it in two steps.