An improved transfrm-finder. Further improvements soon. See msg.
Now, instead of doing a superficial search to try to find the code, we
first perform a deep dive into a block, and see if it is possible to
find what we want there. Only if the deep dive fails, we search for
the narrower path. This matches better with the intuition that most
people will likely have when applying this transformation. Currently,
it still gives up if it is unable to find something via this first
deep dive. The fix to this is to use the evaluation of deep dive
results to "feed back" into further search. At this point, it _should_
become a complete search, which is what we want.