Introduce the check-if-same-printed-code "transformation". See msg.
This "transformation" is a handy debugging tool. It does NOT
transform your code, but simply checks if both the original code
and hint code print to the same value.

This is useful for multiple reasons-
   - Testing if some code you've modified is exactly the same
      (modulo wrapping things in procedures).
   - Sanity-checking the stability of lemmas. The signature of the
      lemma here is the same as other transformations, and thus
      can - be used to perform a first-pass sanity check to ensure
      that - you'll be able to perform other transformations on
      your code.

The concrete reason for introducing this "transformation" is to ensure
that in the process of cleaning up the AESGCM proofs, we don't
accidentally introduce performance regressions. In particular, we
can (locally) check that we can successfully
[check_if_same_printed_code] before performing a full replacement of
the code. This is especially necessary since we will soon be inlining
everything in the "expected code" version.