Make code rely on AESopt (& not AESGCM) for upgraded parts. See msg.
Since we've started working on upgraded versions of proofs for AESGCM,
and this is being done in the [AESopt.vaf] file, it makes sense for us
to start creating a distinction between old and new code.

We do this by removing parts of the code that have already been
upgraded to the new "reorder-ready" style, from AESGCM, and instead
include AESopt into AESGCM.

There are some type-mismatches that I've fixed, but I am performing
all of these changes with the [-codeOnly] flag enabled. Once we have
these stabilized, we should fix up the proofs that broke due to these
changes.