Add convenience module to simplify large debug traces. See msg.
Directly printing from [Vale.X64.Print_s] does not give us full
information about how the [Block (.. (Block ..))] etc wraps around the
code. This makes debugging a tad bit painful when things go
wrong. Also, the amount of code that is dumped is massive. This module
instead sets up overrides for [print_ins] and [print_code] that can be
dropped in to get a shorter/cleaner view on code.

This module is *strictly* meant for debugging purposes, and should not
be included/used anywhere.