Perform a more thorough search to find xform hints. See msg.
Previously, we gave up pretty early if we hit upon a [Block] when
searching for hints. While this was great for most cases, we can
easily hit upon cases where this might be a problem (for example, for
AES256's "inline if" portion). This just means that we need to expand
the search behavior and allow it to look for more unblocked
constructs.

This does _NOT_ change any of the xform hints, but only changes the
search strategy and thus didn't require any proof upgrades :) That
design decision has paid off so damn much! Unfortunately, the process
of debugging the search behavior is still quite painful and requires
poring over tons and tons of trace logs; but at least I don't have to
update proofs each time I update the search behavior to allow for
finding more weird constructs of code.

Hopefully this should be the last of the search upgrades :)