Removed last assume from test_correctness
This is the last unproven assumption from Hacl.Impl.QTesla.fst! Some assumes providing facts from expensive lemmas remain, but these are actually provable, but assumed for now to avoid a whole lot of added verification time during development by disabling the expensive lemmas they require.

These are the assumptions about buffer disjointness and one about modifies. The disjointness ones are really obvious: buffers passed in as arguments to a function are obviously disjoint from buffers created in a local stack frame, but the lemmas needed to prove this take a long time.