Remove all but two assumptions from Hacl.Impl.QTesla.Pack for qTESLA-I
One has to do with guaranteeing coefficients of a secret key fit into the expected range, which I need to verify with Patrick because I don't see anywhere this is validated. The input may just be trusted since it's the secret key.

The other is a pretty obvious inequality relationship that should be easy to prove but is being difficult for some reason. But this is still a significant advance and tests are passing, so making a commit.