Remove more complicated proof from decode_pk for stability's sake
Since we have to assume the output of decode_pk is valid, remove the more
fine-grained proof that was there which would have been used had we been
able to prove the operations actually guaranteed the bounds. Since it turns
out that's not the case, remove all the proof code around slowly expanding
a public key to satisfy the predicate, and just assume the whole thing is
in range with an assumption at the end.

The proof had been assuming the smallest possible fact and then using that
to build to the larger conclusion, but that was making the prover randomly
cancel.