Remove inline_for_extraction from Hacl.Impl.Chacha20.Vec
Now that the tactic is rewriting the call-graph of Impl.Chacha20.Vec,
the wrapper pattern is gone, meaning that inline_for_extraction (in the
Impl file) applies for the entire call-graph. This is very expensive at
extraction-time, and not necessary, since the tactic will automatically
rewrite each function of Impl in the right style, and add the noextract
inline_for_extraction qualifiers automatically.

A corner-case is for leaves of the call-graph, which the tactic does not
re-check: for those, the tactic statically asserts that they have been
marked with inline_for_extraction already, and fails otherwise.

This brings the extraction time of Hacl.Impl.Chacha20.Vec.fst to 42s
(down from 8+ minutes without this fix).