Introduce a basic impl for the peephole transform framework. See msg
The basic idea of this framework is a way to provide a nice
abstraction and make it easier to write new peephole transformers.

Each peephole transformer, given this framework, only needs to write a
simple (list ins -> option (list ins)) function and prove it
"correct", and immediately it obtains a full fledged implementation
that works over general code (including issues wrt control flow
structuring). In particular, this reduces duplication of lots of
common proofs that have nothing to do directly with peephole
optimizations, but instead are to do with preservation of semantics
under an unchanging control flow structure.

The (first) test for such a transformer will be the implementation of
the movbe elimination transformer. This transformer, implemented on
the [jay_movbe-elim-transform] branch is a very simple non-trivial
peephole transformer which required ~450 lines of proof and
implementation combined. The plan is to make this much simpler, by
taking up the burden within this framework.

The other tests for this transformer will be [movbe-intro] (which is a
many->one transformer) and [mov0-xor] (which is a one->one
transformer).

Would be nice to also figure out a plan for a many->many transformer :)