Impl and prove movbe elim correct, under cpuid assumption. See msg.
Previously, when we implemented the movbe elimination
transformation, the semantics of movbe weren't actually sound. This
was because there was no CPUID check on them. Since then, I noticed
that this can cause programs to crash on certain CPUs (which indeed
led to the movbe elimination transform), and since then, the CPUID
check was added.

Due to the addition of this check, we actually need to relax our
statement of correctness of this transformation. In particular, we
need to ensure that the transformation handles this CPUID check
correctly, rather than simply assuming that it is correct. The plan is
to fix this up in some upcoming commits.