Cleaner handling for failing states. See msg.
Previously, we had realized that we didn't need to be as strict as
[equiv_states] to perform the transformation. Thus, we introduced
[equiv_option_states] which places all failing states (either via
returning [None] or by having [ms_ok = false]) into a single bucket.

This works amazingly for the instruction reordering transform, and in
fact _worked_ well for the movbe elimination transform too.

However, with the introduction of the CPUID check (which was a sound
thing to do), proving [equiv_option_states] is now impossible for the
movbe-elimination transform.

However, we do not need to be this strong either. This is because at
the point where the transformer interacts with the Vale code is when
we know that the input code is _definitely_ going to be evaluated to
be "ok", which means that we can relax the requirement for a peephole
transformer to be correct.

The requirement is thus relaxed to be [s1.ok ==> s1 ~ s2], which makes
it easier to prove, and is also sufficient to prove the generic
peephole framework. Additionally, this means we don't need any
assumptions on CPUID for the movbe elimination transformer. Yay! :)