Cherry-pick Jay's commit b5db8ae073a92038a32b9432dfd06fa1142c4d3e
Introduce heaplet indices to locations. See msg.

Some of the proofs that we need to write given the introduction of
these indices _may_ need some abstraction breaking. I am not sure if
this _really_ needs to be done, but at least this sets up the stage
for how these locations should look towards client-facing code, so
that may need to be fixed up.