Fix subtle bug in labeling for unstructured code. See msg.
An implicit assumption being made for the previous version of the code
was invalid, and could potentially lead to bad jump generation. In
trying to explicitly codify this assumption, I realized that the
assumption was false, and instead by restructuring the labelling to
pre-allocate jump labels for unstructured code before generating the
internal block's assembly, would lead to a cleaner design as well as a
correct implementation.