ECDSA Bug in reduction (#331)
This PR solves a problem with missing reduction at the end of ECDSA verification procedure, adds several wycheproof tests for DH and ECDSA and adds one more precondition for DH. Additionally, there was a change in the verification procedure. Before the function 'point_add' there was a check for point equality. If the points are identical, the point_double routine is called.