Update Makefile and dist/configure so that dist/configure is copied in all the target directories containing libintvector.h
Previously, dist/configure was only copied in dist/gcc-compatible. Now
it is copied in more directories, because we now need this script to compute
the proper CFLAGS for the vectorized functions.
Note that the configure script needs the lib_intrinsics.h file to be
present to check if bug 81300 is present or not. If lib_intrinsics.h is
absent, it assumes there is no bug. For now it only assumes there is no
bug when run from dist/mozilla, which is ok because this assumption was
already made before.