Update libintvector.h so that it is possible to tweak the output depending on whether the vectors contain uint32 or uint64 elements
Some generic operations like [and] or [xor], which operate on bits,
don't know whether the vectors contain uint32 or uint64 elements, and
print them as vectors of uint32. This can lead to different traces when
using an architecture with a different endianess. We thus added flags to
take this into account.