Random proof tweaks.
The proof works on HACL CI, but not on Everest CI.
The proof works on OSX, on Linux (interactive) but not on Linux (batch).

The QI profile shows over-triggering of lemma_LowStar.Monotonic.Buffer.loc_disjoint_includes_r

The hints are replayable on OSX and on HACL* CI, but for some reason,
the build is not identical between Everest regression and HACL* CI.

Calling the "right" instances of the lemma does something. Hopefully it
unblocks the CI this time.