Code-quality improvement.
Rewrite split_blocks to ensure that krml's let-match optimization
kicks in (the one that gets rid of a temporary scrutinee of type tuple,
or struct). This removes the need for K___uint32_t_uint32_t in Blake2b,
which in turn eliminates spurious dependency edges between the various
Blake2 variants.