Makefile: introduce resource monitoring with $(RESOURCEMONITOR)
This introduces a RESOURCEMONITOR variable to the Makefile that causes
most rules to generate a .runlim file with information about memory
usage, cpu time, and more. The results can then be tallied with the
`res_summary.sh` script, currently found in the F* repo.

The same idea is applied to the F* and KaRaMeL repos.