Part 2: revert part 1 and do things differently
Using a singleton type, I actually managed to pair up the algorithm with
a non-agile state type (that extracts to C). Using a few tricks related
to erasure (and marking a few more functions erased), I was able to
capture statically the fact that many Keccak streaming functions don't
need the client to pass the algorithm, as the information is either
stored internally, or irrelevant.