EverCrypt.Hash: use @@strict_on_arguments for efficiency
The extraction of EverCrypt.Hash.Incremental has a peak memory usage of
around 11GB, and takes around 10 minutes (on my setup). Some cursory
inspection of F* debug output showed many normalization calls involving
matches to destroy the hash state type, as is done in these functions.

Unfolding this functions in the match is (very) useful if the matched-on
argument is a literal constructor, so the match can reduce, but a waste
that can lead to an exponential explosion otherwise.

Hence, use the strict_on_arguments attribute to make F*
1) eagerly reduce the arguments to these functions before unfolding, and
2) only unfold if the specified arguments are concrete (i.e. the head is
   a constructor).

Re measuring showed that memory usage falls from 11084 MB to 3596 MB
when extracting EverCrypt_Hash_Incremental.krml. Runtime decreases from
691s to 112s (again on my setup, measured in isolation).

A full build passed with this patch, and the `dist` directory ended up
exactly the same on both runs.

There likely are more functions that can be benefit from this attribute.