Explicitly use Z3 4.8.5
F* defaults to using Z3 4.8.5 but we may upgrade this soon to 4.13.3.
Most projects have migrated to this version already. HACL* still has
some proofs that we could not upgrade, so, for the time being, pass a
flag to make F* use Z3 4.8.5.

We plan to keep supporting using F* with 4.8.5, but we'd like the
default version to be newer.

Also, remove `--ext context_pruning` as it is now the default
(it can be disabled with `--ext context_pruning=0`).