Poly1305.Bitvectors: remove a flaky, unused lemma
Bitvectors are primitive in F*, and this proof can only work via Z3
figuring it out. There is no point in a big proof, unless we expand
FStar.BV to expose more facts about bitvectors. That said, Z3 4.8.5
could not figure it out, so this proof was spelled out (though I'm not
sure how it did so, except by randomizing the solver--- there is AFAICT
no lemma about `bv2list` that could help).

Howver, Z3 4.13.3 works pretty reliably. This snippet made the lemma
pass reliably in this file.

	#push-options "--split_queries always"
	let int2bv_uext_64_128 (x : nat) :
	  Lemma  (requires FStar.UInt.size x 64)
		 (ensures  bv_uext #64 #64 (int2bv #64 x) == int2bv #128 x)
	=
	  assert (bv_uext #64 #64 (int2bv #64 x) == int2bv #128 x);
	  ()
	#pop-options

And also works in plain F* (I am adding a test for it FStarLang/FStar#3905).

(Do note however that the assert seems to help... but it does not make a
difference in a clean F* module. I am investigating why separately.)

However, this lemma is unused and does not have an SMT pattern, so this
patch just removes it. We could expose it from the F* library if needed,
perhaps with the sizes being arguments.